RF_09.005.04 Deleted ManageAppearance1Test Tests moved to AppearanceTest
