+++ src/test/java/school/redrover/AppearanceTest.java
@@ -53,6 +53,42 @@ public class AppearanceTest extends BaseTest {
                 "The background color doesn't match the theme");
     }
 
+    @Test
+    public void testDarkThemeApply() {
+        final String actualThemeApplied = new HomePage(getDriver())
+                .clickManageJenkins()
+                .clickAppearanceButton()
+                .clickDarkThemeButton()
+                .clickApplyButton()
+                .getCurrentThemeAttribute();
+
+        Assert.assertEquals(actualThemeApplied, "dark");
+    }
+
+    @Test
+    public void testDefaultThemeApply() {
+        final String actualThemeApplied = new HomePage(getDriver())
+                .clickManageJenkins()
+                .clickAppearanceButton()
+                .clickDefaultThemeButton()
+                .clickApplyButton()
+                .getCurrentThemeAttribute();
+
+        Assert.assertEquals(actualThemeApplied, "none");
+    }
+
+    @Test
+    public void testSystemThemeApply() {
+        final String actualThemeApplied = new HomePage(getDriver())
+                .clickManageJenkins()
+                .clickAppearanceButton()
+                .clickSystemThemeButton()
+                .clickApplyButton()
+                .getCurrentThemeAttribute();
+
+        Assert.assertTrue(actualThemeApplied.contains("system"));
+    }
+
     @AfterMethod
     public void returnToNoneTheme() {
         if (!getDriver().findElement(By.tagName("html")).getAttribute("data-theme").equals("none")) {
